+++ requirements.txt
@@ -16,3 +16,6 @@ flasgger==0.9.7.1
 alembic==1.13.0
 sqlalchemy==2.0.20
 flask-sqlalchemy==3.1.1
+neo4j==5.28.1
+neo4j-graphrag==1.6.1
+neomodel==5.0.0
