+++ src/app/config/dev.py
@@ -2,11 +2,18 @@ import os
 
 DEBUG = os.getenv("DEBUG", "False") == "True"
 
+# MYSQL
 DB_HOST = os.getenv("DB_HOST")  
 DB_PORT = os.getenv("DB_PORT")           
 DB_USER = os.getenv("DB_USER")   
+# NEO4J
+NEO4J_HOST = os.getenv("NEO4J_URL")
+NEO4J_PORT = os.getenv("NEO4J_PORT")
+NEO4J_USER = os.getenv("NEO4J_USER")
+# DB 공통
 DB_PASSWORD = os.getenv("DB_PASSWORD")  
 DB_NAME = os.getenv("DB_NAME")   
+
 GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")  
 
 SWAGGER_SPECS_ROUTE = os.getenv("SWAGGER_SPECS_ROUTE")
