+++ src/app/utils/gemini.py
